fix: get orginial scheduled booking
