+++ packages/features/bookings/lib/handleNewBooking/createBooking.ts
@@ -13,8 +13,8 @@ import type { CalendarEvent } from "@calcom/types/Calendar";
 import type { TgetBookingDataSchema } from "../getBookingDataSchema";
 import type { AwaitedBookingData, EventTypeId } from "./getBookingData";
 import type { NewBookingEventType } from "./getEventTypesFromDB";
-import type { OriginalRescheduledBooking } from "./getOriginalRescheduledBooking";
 import type { LoadedUsers } from "./loadUsers";
+import type { OriginalRescheduledBooking } from "./originalRescheduledBookingUtils";
 import type { PaymentAppData } from "./types";
 
 type ReqBodyWithEnd = TgetBookingDataSchema & { end: string };
